a few different redaction options
